requirements.txt
# Core AI and API packages
google-genai
python-dotenv
elevenlabs
pygame
requests
flask
flask-socketio

# OBS WebSocket integration
obsws-python

# GUI automation and screen capture
pyautogui
pyscreeze
Pillow

# Windows-specific packages
pywin32
winapps

vosk
sounddevice
numpy
opencv-python